refactor(signals): extract shared echarts theme module
Consolidate the duplicated PALETTE, cssHsl, and seriesTotal into
lib/echartsTheme.ts and point QueryChart, PlotChart, ExportDialog,
DebugPage, and SignalTimeChart at it. Removes the forked DebugPage
palette and SignalTimeChart CHART_COLORS so every signals chart shares
one color sequence.
